refactored match rendering methods
