+++ advquery.js
@@ -1,6 +1,6 @@
 var db = require('./db');
-var display = require('./display');
-var computeMatchData = display.computeMatchData;
+var compute = require('./compute');
+var computeMatchData = compute.computeMatchData;
 var utility = require('./utility');
 var isRadiant = utility.isRadiant;
 var mergeObjects = utility.mergeObjects;
